==> HCKT02 - Data Wrangling/1 - file_data.ipynb <==
import pandas as pd
import json
import requests
import glob
import os
from sqlalchemy import create_engine
from bs4 import BeautifulSoup
import time

## Get the Data

### -> Data in files

!ls data/data_0.tsv

def show_dtypes_na(df: pd.DataFrame) -> pd.DataFrame: 
    df_show = pd.concat(
        [
            df.dtypes.sort_index(), 
            df.isna().sum().sort_index()
        ],
        axis=1
    ).rename({0: 'dtypes', 1: 'na'}, axis=1)
    
    return df_show

#### - slice-0: tsv file, compressed 

slice_0 = pd.read_csv('data/data_0.tsv', sep='\t', compression='gzip')
#slice_0.columns = slice_0.columns.str.lower()
slice_0 = slice_0.drop('Unnamed: 0', axis=1)
slice_0 = slice_0.dropna(subset=['Timestamp'])
# id is an int
slice_0.Timestamp = slice_0.Timestamp.astype(int)
slice_0

show_dtypes_na(slice_0)

#### - slice-1: csv with Nones and repeated index

pd.set_option('display.max_columns', 500)

slice_1 = pd.read_csv('data/data_1.csv', sep=',')
# drop rows with no id
slice_1 = slice_1.dropna(subset=['Timestamp'])


slice_1["Timestamp"] = pd.to_datetime(slice_1["Timestamp"]).apply(lambda x: int(time.mktime(x.timetuple())))
# id is an int
slice_1.Timestamp = slice_1.Timestamp.astype(int)
# drop row numbers column
slice_1 = slice_1.drop('Unnamed: 0', axis=1)
slice_1.head(10)

show_dtypes_na(slice_1)

#### - slice-2: JSON file with Nones, values converted to strings

pd.set_option('display.max_columns', 500)

slice_2 = pd.read_csv('data/data_2.csv', sep=',')
# drop rows with no id
slice_2 = slice_2.dropna(subset=['timestamp'])
# id is an int
slice_2.timestamp = slice_2.timestamp.astype(int)
# drop row numbers column
slice_2 = slice_2.drop('Unnamed: 0', axis=1)
slice_2.head(10)

#### slice-3: JSON file with Nones, values converted to strings, multiple lines


slice_3 = pd.read_json('data/data_3.json' ,compression='bz2')
#slice_3 = file_path = os.path.join('data', 'data_3.json')
slice_3['Timestamp'] = slice_3.Timestamp.apply(lambda x: int(time.mktime(x.timetuple())))
slice_3

show_dtypes_na(slice_3)

#### Fix issues on particular dataframes

# change all column names to lower 
slice_0.columns= slice_0.columns.str.lower()
slice_1.columns= slice_1.columns.str.lower()
slice_2.columns= slice_2.columns.str.lower()
slice_3.columns= slice_3.columns.str.lower()


# set the index
slice_0 = slice_0.set_index('timestamp') 
slice_1 = slice_1.set_index('timestamp') 
slice_2 = slice_2.set_index('timestamp') 
slice_3 = slice_3.set_index('timestamp') 

slice_0 = slice_0.dropna(how ='all')
slice_1 = slice_1.dropna(how ='all')
slice_2 = slice_2.dropna(how ='all')
slice_3 = slice_3.dropna(how ='all')


# concatenate dataframes - by default, pd.concat uses an outer join on the index. pd.merge could be an alternative
file_data_df = pd.concat([slice_0, slice_1, slice_2, slice_3]).sort_index()
print("Concatenated shape: ", file_data_df.shape)

# for the same id
# it looks like the values either agree with each other or are null
file_data_df.sort_index()

# so lets fill the null values (within the same column and the same id) 
# with any non null value for the same id
#file_data_df = file_data_df.fillna(method='ffill').fillna(method='bfill').sort_index()
#file_data_df

# then keep only 1 row per id
file_data_df = file_data_df.groupby(file_data_df.index).first()
file_data_df

# check missing data 
file_data_df.isnull().sum()

#file_data_df.dropna()

file_data_df_final=file_data_df.copy()

file_data_df.to_csv('file_data_df_final')

file_data_df_final.sort_index()
