==> HCKT02 - Data Wrangling/2 - database_data.ipynb <==
import pandas as pd
import json
import requests
import glob
import os
from sqlalchemy import create_engine
from bs4 import BeautifulSoup

## Get the Data

table='hackathon_02'
host='batch4-s02-db-instance.ctq2kxc7kx1i.eu-west-1.rds.amazonaws.com'
port=5432
username='ldsa_student'
password='R4Fr4P3aAgMYBqqP'
dbname='s02_db'
schema='public'

conn_str = 'postgresql://{}:{}@{}:{}/{}'.format(username, password, host, port, dbname)
conn_args = {'options': '-csearch_path={}'.format(schema)}

engine = create_engine(conn_str, connect_args=conn_args)

query= 'SELECT * FROM hackathon_02'
df_db_table_info = pd.read_sql_query(query, engine)
display(df_db_table_info)

We have a single table, named "missingdata". We'll be selecting everything from that table and saving it in a dataframe

def show_dtypes_na(df: pd.DataFrame) -> pd.DataFrame: 
    df_show = pd.concat(
        [
            df.dtypes.sort_index(), 
            df.isna().sum().sort_index()
        ],
        axis=1
    ).rename({0: 'dtypes', 1: 'na'}, axis=1)
    
    return df_show

'''query_missingdata = """
SELECT *
FROM hackathon_02
"""
df = pd.read_sql_query(query_missingdata, engine).drop(columns=['index'])
df.columns= df.columns.str.lower()

df'''

query_missingdata = """
SELECT *
FROM hackathon_02
"""
df = pd.read_sql_query(query_missingdata, engine)
df.columns= df.columns.str.lower()

df

show_dtypes_na(df)


df = df[pd.to_numeric(df['timestamp'], errors='coerce').notnull()].set_index('timestamp')
df.index = df.index.astype(int)
df

df.to_csv('database_data.csv')
