==> HCKT02 - Data Wrangling/3 - data_from_html.ipynb <==
import json
# Import the requests library
import requests
# Import some helper functions to print shorter outputs
import utils

# Import some helper packages to handle images
from PIL import Image
from io import BytesIO
from bs4 import BeautifulSoup
import pandas as pd

base_url="https://s02-infrastructure.s3.eu-west-1.amazonaws.com/hackathon-2-batch6"

list_of_df=[]
for i in list(range(0,500)):
    url="/".join([base_url,str(i)])+".html"
    df=pd.read_html(url)[0]
    list_of_df.append(df)

df=pd.concat(list_of_df).reset_index(drop=True)

df.drop(collumns = 'Unamed :0 ', axis =1 )

df.to_csv("data/data_from_html.csv")
