==> HCKT02 - Data Wrangling/4 - api_data.ipynb <==
import pandas as pd
import json
import requests
import glob
import os
from sqlalchemy import create_engine
from bs4 import BeautifulSoup

## Get the Data

### Data in API

base_url = 'https://n6v6uai1jj.execute-api.eu-west-1.amazonaws.com/dev'
response = requests.get(base_url + '/hackathon_02/'+ '1')
print(response.text)

import aiohttp
import asyncio
import time
import pandas as pd

start_time = time.time()


async def get_website(session, url, id):
    async with session.get(url) as resp:
        response = {}
        if resp.status == 200:
            response = await resp.json()
            response['id'] = id
            return response
        return None

async def main():
    responses = {}
    async with aiohttp.ClientSession() as session:

        tasks = []
        for number in range(1, 20000):
            url = f'https://n6v6uai1jj.execute-api.eu-west-1.amazonaws.com/dev/hackathon_02/{number}'
            tasks.append(asyncio.ensure_future(get_website(session, url, number)))

        websites = await asyncio.gather(*tasks)
        for website in websites:
            if website:
                responses[website['id']] = website
    
    df = pd.DataFrame.from_dict(responses, orient='index')
    df.to_csv('raw_api_data.csv', index=False)

await main()
print("--- %s seconds ---" % (time.time() - start_time))

! head raw_api_data.csv

content = pd.read_csv('raw_api_data.csv', index_col=0)

content

content.columns= content.columns.str.lower()
content = content[pd.to_numeric(content['id'], errors='coerce').notnull()]
content.index = content.index.astype(int)
content  = content.drop('id', axis =1)
content

content.to_csv("api_data.csv")
