==> HCKT02 - Data Wrangling/5 - merge_data.ipynb <==
import pandas as pd
import json
import requests
import glob
import os
from sqlalchemy import create_engine
from bs4 import BeautifulSoup


## Merge the Data

We'll be concatenating all the dataframes, droping any duplicated entries and observations with missing data

pd.read_csv('file_data_df_final.csv', index_col='timestamp').sort_index().iloc[[-1]].index

pd.read_csv('database_data.csv', index_col='timestamp').sort_index().iloc[[-1]].index

#html= pd.read_csv('data_from_html.csv', index_col='Timestamp').sort_index().iloc[[-1]].index
html= pd.read_csv('data_from_html.csv', index_col='Timestamp')
html = html.drop(columns =["Unnamed: 0.1",'Unnamed: 0'], axis =1)
html.columns = html.columns.str.lower()

html.to_csv('website_data.csv')

pd.read_csv('api_data.csv', index_col='Timestamp')

df = pd.concat([
    pd.read_csv('api_data.csv', index_col='Timestamp'),
    pd.read_csv('file_data_df_final.csv', index_col='timestamp'),
    pd.read_csv('database_data.csv', index_col='timestamp'),
    pd.read_csv('website_data.csv', index_col='Timestamp'),
    pd.read_csv('signal.csv', index_col='timestamp')
    
])
df.sort_index()

# so lets fill the null values (within the same column and the same id) 
# with any non null value for the same id
#df = df.fillna(method='ffill').fillna(method='bfill').sort_index()


# then keep only 1 row per id
df = df.groupby(df.index).first()
#df = df.drop(columns =['Unnamed: 0'], axis =1)
df.index.name = 'timestamp'

#df=df.dropna(how = 'all')
#df = df.dropna(subset = 'signal')
#df

## Create Train Dataset

df.to_csv("merge_data.csv")
